requirements.txt
black==24.4.2
boto3==1.34.148
cachecontrol==0.14.0
django-cors-headers==4.4.0
django-storages==1.14.4
google-auth==2.32.0
googleapis-common-protos==1.63.2
grpcio==1.65.1
h11==0.14.0
httplib2==0.22.0
isort==5.13.2
minio==7.2.7
pillow==10.4.0
pip-chill==1.0.3
proto-plus==1.24.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-dotenv==1.0.1
pywebpush==2.0.0
redis==5.0.7
swapper==1.3.0
uritemplate==4.1.1
whitenoise==6.7.0
